+++ clickable/libconfig.py
@@ -31,7 +31,10 @@ class LibConfig(object):
 
     install = False
 
-    def __init__(self, json_config):
+    def __init__(self, name, json_config, debug_build):
+        self.debug_build = debug_build
+        self.name = name
+
         self.config = {
             'postmake': None,
             'prebuild': None,
@@ -67,8 +70,8 @@ class LibConfig(object):
             super().__setattr__(name, value)
 
     def set_dirs(self):
-        self.config['dir'] = os.path.join(self.cwd, self.config['dir']) if self.config['dir'] else os.path.join(self.cwd, 'build', self.config['name'])
-        self.config['src_dir'] = os.path.join(self.cwd, self.config['src_dir']) if self.config['src_dir'] else os.path.join(self.cwd, 'libs', self.config['name'])
+        self.config['dir'] = os.path.join(self.cwd, self.config['dir']) if self.config['dir'] else os.path.join(self.cwd, 'build', self.name)
+        self.config['src_dir'] = os.path.join(self.cwd, self.config['src_dir']) if self.config['src_dir'] else os.path.join(self.cwd, 'libs', self.name)
         self.temp = self.config['dir']
 
     def cleanup_config(self):